+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Plutus/TxInfo.hs
@@ -226,6 +226,7 @@ toLegacyPlutusArgs proxy pv sh mkScriptContext scriptPurpose maybeSpendingData r
 
 instance EraPlutusContext AlonzoEra where
   type ContextError AlonzoEra = AlonzoContextError AlonzoEra
+  type LevelTxInfo level AlonzoEra = ()
   newtype TxInfoResult AlonzoEra
     = AlonzoTxInfoResult (PlutusTxInfoResult 'PlutusV1 AlonzoEra)
 
@@ -240,6 +241,8 @@ instance EraPlutusContext AlonzoEra where
 
   lookupTxInfoResult SPlutusV1 (AlonzoTxInfoResult tirPlutusV1) = tirPlutusV1
   lookupTxInfoResult slang _ = lookupTxInfoResultImpossible slang
+
+  mkTopTxInfo = ()
 
 data AlonzoContextError era
   = TranslationLogicMissingInput TxIn