+++ eras/alonzo/impl/testlib/Test/Cardano/Ledger/Alonzo/Translation/TranslatableGen.hs
@@ -17,7 +17,7 @@ module Test.Cardano.Ledger.Alonzo.Translation.TranslatableGen (
 
 import Cardano.Ledger.Alonzo (AlonzoEra)
 import Cardano.Ledger.Alonzo.Plutus.Context (
-  EraPlutusContext,
+  EraPlutusContext (..),
   EraPlutusTxInfo (..),
   LedgerTxInfo (..),
   PlutusTxInfo,
@@ -93,7 +93,7 @@ genTranslationInstance = do
           , ltiSystemStart = systemStart
           , ltiUTxO = utxo
           , ltiTx = tx
-          , ltiMemoizedSubTransactions = mempty
+          , ltiLevelInfo = mkTopTxInfo @era
           }
   plutusPurpose <- arbitrary
   pure $ case supportedLanguage of