+++ eras/dijkstra/impl/CHANGELOG.md
@@ -30,6 +30,9 @@
   - `duePParams` -> `uePParams`
   - `dueCertState` -> `ueOriginalCertState`
   - `dueOriginalUtxo` -> `ueOriginalUtxo`
+* Add `Default (LevelTxInfo TopTx era)` and `LevelTxInfo SubTx era ~ DijkstraLevelTxInfo SubTx era` constraints to `mkDijkstraStAnnTopTx`
+* Add `LevelTxInfo SubTx era ~ DijkstraLevelTxInfo SubTx era` constraint to `mkDijkstraStAnnSubTx`
+* Add `DijkstraLevelTxInfo` as the `LevelTxInfo` instance
 * Add `ScriptHashNotFoundForPurpose` constructor to `DijkstraContextError`
 * Change `PointerPresentInOutput` constructor of `DijkstraContextError` to contain a `NonEmptySet TxOutSource` instead of `NonEmpty (TxOut era)`
 * Add `udppPlutusV4CostModel` field to `UpgradeDijkstraPParams`
@@ -73,6 +76,7 @@
 * Remove the `invalid_transactions` rule and drop the field from `block_body`
 * Add `HuddleRule "vrf_cert"` instance
 * Add Leios protocol parameter entries (tags 40-48) in `protocol_param_update`
+* Add `Default (LevelTxInfo TopTx era)` and `LevelTxInfo SubTx era ~ DijkstraLevelTxInfo SubTx era` constraints to some testing functions
 * Add `max_pledge_leverage` rule and its entry in `protocol_param_update`
 
 ### `testlib`